Move wizard and steps into parent module
Allow wizard to handle participant and cohort support query outcomes, this depends on the presence of participant_id.
