+++ app/views/schools/early_career_teachers/change_lead_provider/check_your_answers.html.erb
@@ -29,7 +29,7 @@
           row.with_action(
             text: "Change",
             visually_hidden_text: "email address",
-            href: schools_early_career_teachers_change_lead_provider_email_path(default_path_params)
+            href: schools_change_lead_provider_email_path(default_path_params)
           )
         end
         sl.with_row do |row|
@@ -42,7 +42,7 @@
           row.with_action(
             text: "Change",
             visually_hidden_text: "lead provider",
-            href: schools_early_career_teachers_change_lead_provider_lead_provider_path(default_path_params)
+            href: schools_change_lead_provider_lead_provider_path(default_path_params)
           )
         end
       end %>
