+++ app/wizards/schools/change_lead_provider/check_your_answers_step.rb
@@ -1,0 +1,21 @@
+# frozen_string_literal: true
+
+module Schools
+  module ChangeLeadProvider
+    class CheckYourAnswersStep < StoredStep
+      attr_accessor :complete
+
+      def self.permitted_params
+        [:complete]
+      end
+
+      def previous_step
+        :lead_provider
+      end
+
+      def next_step
+        :success
+      end
+    end
+  end
+end
