+++ app/wizards/schools/change_lead_provider/contact_providers_step.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+class Schools::ChangeLeadProvider::ContactProvidersStep < DfE::Wizard::Step; end
