--- app/wizards/schools/early_career_teachers/change_lead_provider/contact_providers_step.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-class Schools::EarlyCareerTeachers::ChangeLeadProvider::ContactProvidersStep < DfE::Wizard::Step; end
