--- app/wizards/schools/early_career_teachers/change_lead_provider/intro_step.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-class Schools::EarlyCareerTeachers::ChangeLeadProvider::IntroStep < DfE::Wizard::Step; end
